build[ui] Fix yarn test failure, update test
